requirements.txt
pulumi_docker==4.5.4
python-dotenv==1.0.1

